moving the test code to main.cpp
